@@ -1,21 +1,19 @@
 [package]
 name = "corgi"
-version = "0.1.0"
-edition = "2021"
+version = "0.2.0"
+edition = "2024"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-bytemuck = { version = "1.13", features = ["derive"] }
+bytemuck = { version = "1.23", features = ["derive"] }
 color-eyre = "0.6"
-egui = "0.21"
-egui-wgpu = { version = "0.21", features = ["winit"] }
-egui-winit = "0.21"
-image = "0.24"
-rayon = "1.7"
-rug = "1.19"
+eframe = { version = "0.32.0", features = ["wgpu"] }
+image = "0.25"
+rayon = "1.10"
+rug = "1.27"
 tokio = { version = "1", features = ["macros", "rt-multi-thread", "sync"] }
 tracing = "0.1"
 tracing-subscriber = "0.3"
-wgpu = "0.15"
-winit = "0.28"
+wgpu = "25"
+
